Get drug exposure entries for adherence calculations

Get drug exposure entries for adherence calculations

Usage

generateChronicDrugExposure(
  cdm,
  name = "adherenceFromOMOP",
  indexDateForPatientCalculation = "drug_exposure_start_date",
  overwrite = FALSE,
  subjects = NULL,
  drugConceptIdList = NULL,
  cohortSchema = NULL,
  cohortId = NULL,
  cohortTableName = "cohort",
  ...
)

Arguments

cdm

A cdm reference via CDMConnector.

name

Name of the new cohort table, it must be a length 1 character vector.

indexDateForPatientCalculation

Value for indexDate in addDeathDays function from PatientProfiles.

overwrite

If table with provided name exists this value chooses the action.

subjects

Possibility to choose a specific set of subjects.

drugConceptIdList

list with drug concept ids that determine if drug exposure records are included or not

cohortSchema

Possible include only certain people that exist in a specific defined cohort between a time period. Parameter cohort schema must be a string value of the database schema with the defined cohort.

cohortId

List of numeric values of the defined cohorts that haveS included person ids as well as start and end date of the cohort.

cohortTableName

usually named "cohort"

...

Arguments to be passed to CodelistGenerator getDrugIngredientCodes function.

Value

reference to table

Examples

if (FALSE) { # \dontrun{
cdm <- mockAdherenceFromOMOP()

chronicDrugExposure <- generateChronicDrugExposure(
  cdm = cdm,
  drugConceptIdList = concepts_ids,
  name = "chronic_drug_exposure_table",
  overwrite = T
)
} # }