Generate a table in the cdm containing drug exposure events with the specified ingredients

Generate a table in the cdm containing drug exposure events with the specified ingredients

generateIngredientSet(
  cdm,
  name = "adherenceFromOMOP",
  ingredients,
  cohortSchema = NULL,
  cohortId = NULL
)

Arguments

cdm

A cdm_reference object

name

Name of the new cohort table, it must be a length 1 character vector.

ingredients

Specifies medication under observation. Uses RxNorm Ingredient concept code as acceptable parameter.

cohortSchema

Possible include only certain people that exist in a specific defined cohort between a time period. Parameter cohort schema must be a string value of the database schema with the defined cohort.

cohortId

Numeric value of the defined cohort that has included person ids as well as start and end date of the cohort.

Value

The function returns the cdm reference with a new table